Move database URL parsing into Electric.Config
