+++ components/electric/lib/electric/utils.ex
@@ -195,155 +195,6 @@ defmodule Electric.Utils do
   defp c(_), do: throw(:error)
 
   @doc """
-  Parse a PostgreSQL URI into a keyword list.
-
-  ## Examples
-
-      iex> parse_postgresql_uri("postgresql://postgres:password@example.com/app-db")
-      [
-        host: "example.com",
-        port: 5432,
-        database: "app-db",
-        username: "postgres",
-        password: "password",
-      ]
-
-      iex> parse_postgresql_uri("postgresql://electric@192.168.111.33:81/__shadow")
-      [
-        host: "192.168.111.33",
-        port: 81,
-        database: "__shadow",
-        username: "electric"
-      ]
-
-      iex> parse_postgresql_uri("postgresql://pg@[2001:db8::1234]:4321")
-      [
-        host: "2001:db8::1234",
-        port: 4321,
-        database: "pg",
-        username: "pg"
-      ]
-
-      iex> parse_postgresql_uri("postgresql://user@localhost:5433/")
-      [
-        host: "localhost",
-        port: 5433,
-        database: "user",
-        username: "user"
-      ]
-
-      iex> parse_postgresql_uri("postgresql://localhost")
-      ** (RuntimeError) Invalid or missing username in DATABASE_URL
-
-      iex> parse_postgresql_uri("postgresql://:@localhost")
-      ** (RuntimeError) Invalid or missing username in DATABASE_URL
-
-      iex> parse_postgresql_uri("postgresql://:password@localhost")
-      ** (RuntimeError) Invalid or missing username in DATABASE_URL
-
-      iex> parse_postgresql_uri("postgresql://user:password")
-      ** (RuntimeError) Invalid or missing username in DATABASE_URL
-
-      iex> parse_postgresql_uri("postgresql://user:password@")
-      ** (RuntimeError) Missing host in DATABASE_URL
-
-      iex> parse_postgresql_uri("postgresql://user@localhost:5433/mydb?options=-c%20synchronous_commit%3Doff")
-      ** (RuntimeError) Electric does not support any query options in DATABASE_URL.
-
-      iex> parse_postgresql_uri("postgresql://electric@localhost/db?replication=database")
-      ** (RuntimeError) Electric does not support the "replication" option. It opens both a replication connection and regular connections to Postgres as needed.
-
-      iex> parse_postgresql_uri("postgresql://electric@localhost/db?replication=off")
-      ** (RuntimeError) Electric does not support the "replication" option. It opens both a replication connection and regular connections to Postgres as needed.
-
-      iex> parse_postgresql_uri("postgres://super_user@localhost:7801/postgres?sslmode=yesplease")
-      ** (RuntimeError) Electric does not support the "sslmode" option. Use the DATABASE_REQUIRE_SSL configuration option instead.
-  """
-  @spec parse_postgresql_uri(binary) :: keyword
-  def parse_postgresql_uri(uri_str) do
-    %URI{scheme: scheme, host: host, port: port, path: path, userinfo: userinfo, query: query} =
-      URI.parse(uri_str)
-
-    :ok = assert_valid_scheme!(scheme)
-
-    :ok = assert_valid_host!(host)
-    port = port || 5432
-
-    {username, password} = parse_userinfo!(userinfo)
-
-    database = parse_database(path, username)
-
-    query_params =
-      if query do
-        URI.decode_query(query)
-      else
-        %{}
-      end
-
-    :ok = assert_no_query_params(query_params)
-
-    {:ok,
-     [
-       host: host,
-       port: port,
-       database: database,
-       username: username,
-       password: password
-     ]
-     |> Enum.reject(fn {_key, val} -> is_nil(val) end)}
-  end
-
-  defp assert_no_query_params(params) when map_size(params) == 0, do: :ok
-
-  defp assert_no_query_params(%{"sslmode" => _}) do
-    raise "Electric does not support the \"sslmode\" option. Use the DATABASE_REQUIRE_SSL configuration option instead."
-  end
-
-  defp assert_no_query_params(%{"replication" => _}) do
-    raise "Electric does not support the \"replication\" option. It opens both a replication connection and regular connections to Postgres as needed."
-  end
-
-  defp assert_no_query_params(_) do
-    raise "Electric does not support any query options in DATABASE_URL."
-  end
-
-  defp assert_valid_scheme!(scheme) when scheme in ["postgres", "postgresql"], do: :ok
-
-  defp assert_valid_scheme!(scheme) do
-    raise "Invalid scheme in DATABASE_URL: #{inspect(scheme)}"
-  end
-
-  defp assert_valid_host!(str) do
-    if is_binary(str) and String.trim(str) != "" do
-      :ok
-    else
-      raise "Missing host in DATABASE_URL"
-    end
-  end
-
-  defp parse_userinfo!(str) do
-    try do
-      true = is_binary(str)
-
-      {username, password} =
-        case String.split(str, ":") do
-          [username] -> {username, nil}
-          [username, password] -> {username, password}
-        end
-
-      false = String.trim(username) == ""
-
-      {username, password}
-    rescue
-      _ -> raise "Invalid or missing username in DATABASE_URL"
-    end
-  end
-
-  defp parse_database(nil, username), do: username
-  defp parse_database("/", username), do: username
-  defp parse_database("/" <> dbname, _username), do: dbname
-
-  @doc """
   Output a 2-tuple relation (table) reference as pg-style `"schema"."table"`.
   """
   @spec inspect_relation({String.t(), String.t()}) :: String.t()
