+++ components/electric/test/electric/config_test.exs
@@ -1,0 +1,4 @@
+defmodule Electric.ConfigTest do
+  use ExUnit.Case, async: true
+  doctest Electric.Config, import: true
+end
