+++ components/electric/test/electric/utils_test.exs
@@ -1,4 +1,4 @@
-defmodule Electric.Test.UtilsTest do
+defmodule Electric.UtilsTest do
   alias Electric.Utils
   use ExUnit.Case, async: true
   doctest Utils, import: true
